+++ packages/rocketh-deploy/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @rocketh/deploy
 
+## 0.11.22-testing.5
+
+### Patch Changes
+
+- allow to pass Extra date to environment
+
 ## 0.11.22-testing.4
 
 ### Patch Changes