+++ packages/rocketh-proxy/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @rocketh/proxy
 
+## 0.11.26-testing.5
+
+### Patch Changes
+
+- allow to pass Extra date to environment
+- Updated dependencies
+  - @rocketh/read-execute@0.11.23-testing.5
+  - @rocketh/deploy@0.11.22-testing.5
+
 ## 0.11.26-testing.4
 
 ### Patch Changes